feat: refactor librosa compatibility test
